requirements.txt
flask
boto3
pandas
python-dotenv
os